// src/api-client/models/BaseCommissioningTest.ts
export interface BaseCommissioningTest {
  /**
   *
   * @type {string}
   * @memberof BaseCommissioningTest
   */
  uuid: string;
  /**
   *
   * @type {string}
   * @memberof BaseCommissioningTest
   */
  status: BaseCommissioningTestStatusEnum;
  /**
   *
   * @type {Date}
   * @memberof BaseCommissioningTest
   */
  createdAt: Date;
  /**
   *
   * @type {Date}
   * @memberof BaseCommissioningTest
   */
  startedAt: Date | null;
  /**
   *
   * @type {Date}
   * @memberof BaseCommissioningTest
   */
  updatedAt: Date;
  /**
   *
   * @type {Date}
   * @memberof BaseCommissioningTest
   */
  endedAt: Date | null;
  /**
   *
   * @type {boolean}
   * @memberof BaseCommissioningTest
   */
  forced: boolean;
  /**
   *
   * @type {string}
   * @memberof BaseCommissioningTest
   */
  commissioningUuid: string;
  /**
   *
   * @type {CommissioningTestType}
   * @memberof BaseCommissioningTest
   */
  type: CommissioningTestType;
  /**
   *
   * @type {string}
   * @memberof BaseCommissioningTest
   */
  installerUuid: string;
  /**
   *
   * @type {number}
   * @memberof BaseCommissioningTest
   */
  step: number;
  /**
   *
   * @type {Date}
   * @memberof BaseCommissioningTest
   */
  maximumDatetime: Date;
}

/**
 * @export
 */
export const BaseCommissioningTestStatusEnum = {
  Success: "SUCCESS",
  Failed: "FAILED",
  InProgress: "IN_PROGRESS",
  Cancelled: "CANCELLED",
  Waiting: "WAITING",
  Ready: "READY",
  Validated: "VALIDATED",
} as const;
export type BaseCommissioningTestStatusEnum =
  (typeof BaseCommissioningTestStatusEnum)[keyof typeof BaseCommissioningTestStatusEnum];